requirements.txt
django==4.0
djangorestframework==3.12.4
pytz==2021.3



black==21.7b0
isort==5.10.1
autoflake==1.4
